rename Chats component to Pins
